Update README.md
Ever think of penning a juicy tale of mystery in an exotic locale.. starring your favorite pet with smarts-&-attitude? You're in the right place! Although it may be a humble command-line generator at the moment, expect this baby to be fully spun out into one of the awesome-est full-stack apps you've seen in a while! Who says Choose-Your-Own-Adventure is past its prime?!